+++ src/xcfg/__init__.py
@@ -40,7 +40,7 @@ from .loader import (
 )
 from .spec import DEFAULT_CONFIG_NAME, ENV_CONFIG_PREFIX, ConfigSpec, RawConfig
 
-__version__ = "0.3.1"
+__version__ = "0.4.0"
 
 __all__ = [
     "DEFAULT_CONFIG_NAME",